Revamped Game interface, added Game.Default. Added PlayN.tick().
TL;DR: change "implements Game" to "extends Game.Default", delete your
updateRate() implementation and instead pass that value to Game.Default's
constructor.

Details:

- PlayN.tick() returns the number of ms that have elapsed since game startup
  (as an int, which is safe to use in JavaScript or otherwise).

- Game.update/paint/updateRate went away (sort of) to be replaced by Game.tick.
  A game that wants to handle updating and painting via some custom mechanism
  now just implements Game.tick and does their own simulation updating and
  scene graph interoplation.

  "Normal" games now extend Game.Default instead of implementing Game.
  Game.Default provides update(int delta) (which eliminates a pointless and
  problematic use of floats) and paint(float alpha). Game.updateRate goes away
  and is passed to Game.Default's constructor. You can no longer change your
  game's update rate on the fly, which was wacky.

  Game.Default implements the update/paint approach that was previously (sort
  of) mandated for all games. There is no longer the "updateRate = 0" hackery
  because games that want to do custom things just override tick() and don't
  have to second guess what the platform is doing.

  Game.Default.update differs from the old Game.update in that updates will be
  coalesced if the game is running slowly. See Game.Default.update's javadocs
  for details.

  Game.Default.paint differs from the old Game.paint in that alpha can be
  greater than one if your game is running slowly (but this is really the right
  thing to do and your code will probably just do the right thing if you're
  doing interpolation math normally).

  Game.Default's calculation of the 'alpha' value passed to paint is more
  accurate and should result in smoother display in games that actually use it
  to interpolate between simulation updates.

+++ tests/core/src/main/java/playn/tests/core/ClippedGroupTest.java
@@ -118,8 +118,8 @@ public class ClippedGroupTest extends Test {
   }
 
   @Override
-  public void update(float delta) {
-    elapsed += delta/1000;
+  public void update(int delta) {
+    elapsed += delta/1000f;
     i1.setRotation(elapsed * FloatMath.PI/2);
     s1.setRotation(elapsed * FloatMath.PI/2);
     g2.setWidth(Math.max(1f, Math.abs(100 * FloatMath.sin(elapsed))));
